refactor: token warnings in one hook
